requirements.txt
streamlit
joblib
nltk
sastrawi
pandas
Pillow
